requirements.txt
Flask==2.3.2
Flask_MySQLdb==1.0.1
mysql_connector_repackaged==0.3.1
pytz==2023.3
tqdm==4.65.0
